[WIP] Started Unit Tests
